refactor(marketing): rename router path for showcase pages + rename file features-side-image-section
