+++ apps/marketing/src/pages/landing/sections/hero-simple-section.tsx
