+++ apps/marketing/src/pages/showcase/blog-card.tsx
