+++ apps/marketing/src/pages/showcase/features-side-image-section.tsx
